types fixes
