+++ ui/components/Button/Button.tsx
@@ -1,32 +1,10 @@
 import React from "react";
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 import { AllContentProps, ButtonProps, SizeProps, ViewProps } from "./types";
 import { ITheme } from "../../themes";
+import { sizes } from "./buttonViews";
 
-export const sizes = {
-    l: {
-        padding: "20px",
-        fontSize: "20px",
-    },
-    m: {
-        padding: "15px",
-        fontSize: "15px",
-    },
-    s: {
-        padding: "10px",
-        fontSize: "12px",
-    },
-};
-
-const applySizes = ({ size = "m" }: SizeProps) => {
-    return css`
-        padding: ${sizes[size].padding};
-        font-size: ${sizes[size].fontSize};
-    })}
-    `;
-};
-
-export const StyledButton = styled.button<ITheme & ViewProps>`
+export const StyledButton = styled.button<ITheme & ViewProps & SizeProps>`
     position: relative;
     display: inline-flex;
     align-items: center;
@@ -35,17 +13,16 @@ export const StyledButton = styled.button<ITheme & ViewProps>`
     border: none;
     cursor: pointer;
     border-radius: 10px;
-    color: ${({ theme, view }) => theme.colors.buttons[view!].color};
-    background: ${({ theme, view }) =>
-        theme.colors.buttons[view!].backgroundColor};
+    color: ${({ theme, view = "primary" }) => theme.colors.buttons[view].color};
+    background: ${({ theme, view = "primary" }) =>
+        theme.colors.buttons[view].backgroundColor};
 
     &:focus {
         outline: none;
     }
-    ${
-        // @ts-ignore
-        applySizes
-    }
+
+    padding: ${({ size = "s" }) => sizes[size].padding};
+    font-size: ${({ size = "s" }) => sizes[size].fontSize};
 `;
 
 export const Button = React.forwardRef<
@@ -57,8 +34,6 @@ export const Button = React.forwardRef<
     return (
         <StyledButton
             ref={ref as React.MutableRefObject<HTMLButtonElement>}
-            view="primary"
-            size="s"
             {...rest}
         >
             {children}
